changes 29.10.20

+++ 2.2.2. spring_mvc/src/main/java/web/dao/UserDao.java
@@ -15,7 +15,7 @@ public interface UserDao {
 
     List<User> getAllUsers() throws Exception;
 
-    User getUserById(int id) throws  Exception;
+    User getUserById(long id) throws  Exception;
 
     boolean checkLastName(String lastName)throws Exception;
