+++ 2.2.2. spring_mvc/src/main/java/web/service/UserService.java
@@ -13,7 +13,7 @@ public interface UserService {
     void removeUser(User user) throws Exception;
     void updateUser(User user)throws Exception;
     List<User> getAllUsers()throws Exception;
-    User getUserById(int id)throws Exception;
+    User getUserById(long id)throws Exception;
     boolean checkLastName(String lastName)throws Exception;
 
 }
